Rename the ConsoleCommandListener class to ConsoleListener
